requirements.txt
numpy==1.19.5
packaging==21.3
Pillow==9.3.0
scikit_learn==1.1.3
seqeval==1.2.2
tensorflow==2.10.0
tensorflow_gpu==2.4.0
torch==1.7.1+cu110
TorchCRF==1.1.0
torchvision==0.8.2+cu110
tqdm==4.64.0
transformers==4.11.3
